维护数据功能-317

+++ main/java/com/liaoyuan/web/entity/DataBean.java
@@ -13,7 +13,7 @@ public class DataBean {
     private int pId;
 
     //吨数
-    private int tonnage;
+    private double tonnage;
 
     //运费
     private double cost;
